+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.13.1 (unreleased)
 -------------------
 
+- Add CSS / JS for minisite dropdown menu : WEBOTT-9
+  [laulaz]
+
 - Fix viewlets.xml that didn't apply at all
   [laulaz]